+++ CHANGELOG.md
@@ -483,6 +483,8 @@
 ## [unreleased]
 
 - Relegate "Download report as CSV" link to tertiary status
+- Add a `funding_type` column to a budget
+- Work out a Budget's period based on the financial year
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-30...HEAD
 [release-30]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-29...release-30